dispatcher/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
google-auth>=2.33,<3
requests>=2.32,<3
structlog>=24.1.0
